+++ docs/changelog.rst
@@ -23,6 +23,8 @@ In progress (unreleased)
 
 * :gh:issue:`1529` :mod:`ansible_mitogen`: Make injected ``__main__`` on target
   more identifiable/greppable
+* :gh:issue:`1529` :mod:`ansible_mitogen`: Remove
+  :attr:`ansible_mitogen.runner.NewStyleRunner.main_module_name`
 * :gh:issue:`1523` tests: Bump Ansible 14 tests to 14.x